wpt_babel/content-security-policy/nonce-hiding/script-nonces-hidden.html

<!DOCTYPE html>
<script nonce="abc" src="/resources/testharness.js">
</script>
<script nonce="abc" src="/resources/testharnessreport.js">
</script>
<!-- `Content-Security-Policy: script-src 'nonce-abc'; img-src 'none'` delivered via headers -->
<body>
 <!-- Basics -->
 <script id="testScript" nonce="abc">
  "use strict";

document.currentScript.setAttribute('executed', 'yay');
 </script>
 <script nonce="abc">
  "use strict";

var script = document.querySelector('#testScript');
test(function (t) {
  // Query Selector
  assert_equals(document.querySelector('body [nonce]'), script);
  assert_equals(document.querySelector('body [nonce=""]'), script);
  assert_equals(document.querySelector('body [nonce=abc]'), null);
  assert_equals(script.getAttribute('nonce'), '');
  assert_equals(script.nonce, 'abc');
}, "Reading 'nonce' content attribute and IDL attribute.");

// Clone node.
test(function (t) {
  script.setAttribute('executed', 'boo');
  var s2 = script.cloneNode();
  assert_equals(s2.nonce, 'abc', 'IDL attribute');
  assert_equals(s2.getAttribute('nonce'), '');
}, "Cloned node retains nonce.");
async_test(function (t) {
  var s2 = script.cloneNode();
  document.head.appendChild(s2);
  assert_equals(s2.nonce, 'abc');
  assert_equals(s2.getAttribute('nonce'), '');
  window.addEventListener('load', t.step_func_done(function (_) {
    // The cloned script won't execute, as its 'already started' flag is set.
    assert_equals(s2.getAttribute('executed'), 'boo');
  }));
}, "Cloned node retains nonce when inserted.");

// Set the content attribute to 'foo'
test(function (t) {
  script.setAttribute('nonce', 'foo');
  assert_equals(script.getAttribute('nonce'), 'foo');
  assert_equals(script.nonce, 'foo');
}, "Writing 'nonce' content attribute.");

// Set the IDL attribute to 'bar'
test(function (t) {
  script.nonce = 'bar';
  assert_equals(script.nonce, 'bar');
  assert_equals(script.getAttribute('nonce'), 'foo');
}, "Writing 'nonce' IDL attribute.");

// Fragment parser.
var documentWriteTest = async_test("Document-written script executes.");
document.write("<script nonce='abc'>\n      documentWriteTest.done();\n      test(t => {\n        var script = document.currentScript;\n        assert_equals(script.getAttribute('nonce'), '');\n        assert_equals(script.nonce, 'abc');\n      }, \"Document-written script's nonce value.\");\n    </scr" + "ipt>");

// Create node.
async_test(function (t) {
  var s = document.createElement('script');
  s.innerText = script.innerText;
  s.nonce = 'abc';
  assert_equals(s.nonce, 'abc');
  assert_equals(s.getAttribute('nonce'), null);
  document.head.appendChild(s);
  assert_equals(s.nonce, 'abc');
  assert_equals(s.getAttribute('nonce'), null);
  window.addEventListener('load', t.step_func_done(function (_) {
    assert_equals(s.getAttribute('executed'), 'yay');
  }));
}, "createElement.nonce.");
async_test(function (t) {
  var s = document.createElement('script');
  s.innerText = script.innerText;
  s.nonce = 'zyx';
  s.setAttribute('nonce', 'abc');
  assert_equals(s.nonce, 'abc');
  document.head.appendChild(s);
  assert_equals(s.nonce, 'abc');
  assert_equals(s.getAttribute('nonce'), '');
  window.addEventListener('load', t.step_func_done(function (_) {
    assert_equals(s.getAttribute('executed'), 'yay');
  }));
}, "setAttribute('nonce') overwrites '.nonce' upon insertion.");

// Create node.
async_test(function (t) {
  var s = document.createElement('script');
  s.innerText = script.innerText;
  s.setAttribute('nonce', 'abc');
  assert_equals(s.getAttribute('nonce'), 'abc', "Pre-insertion content");
  assert_equals(s.nonce, 'abc', "Pre-insertion IDL");
  document.head.appendChild(s);
  assert_equals(s.nonce, 'abc', "Post-insertion IDL");
  assert_equals(s.getAttribute('nonce'), '', "Post-insertion content");
  window.addEventListener('load', t.step_func_done(function (_) {
    assert_equals(s.getAttribute('executed'), 'yay');
  }));
}, "createElement.setAttribute.");
 </script>
 <!-- Custom Element -->
 <script nonce="abc">
  "use strict";

function _typeof(o) { "@babel/helpers - typeof"; return _typeof = "function" == typeof Symbol && "symbol" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && "function" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? "symbol" : typeof o; }, _typeof(o); }
function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }
function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, "prototype", { writable: false }); return Constructor; }
function _toPropertyKey(t) { var i = _toPrimitive(t, "string"); return "symbol" == _typeof(i) ? i : String(i); }
function _toPrimitive(t, r) { if ("object" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || "default"); if ("object" != _typeof(i)) return i; throw new TypeError("@@toPrimitive must return a primitive value."); } return ("string" === r ? String : Number)(t); }
function _callSuper(t, o, e) { return o = _getPrototypeOf(o), _possibleConstructorReturn(t, _isNativeReflectConstruct() ? Reflect.construct(o, e || [], _getPrototypeOf(t).constructor) : o.apply(t, e)); }
function _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === "object" || typeof call === "function")) { return call; } else if (call !== void 0) { throw new TypeError("Derived constructors may only return object or undefined"); } return _assertThisInitialized(self); }
function _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return self; }
function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, "prototype", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }
function _wrapNativeSuper(Class) { var _cache = typeof Map === "function" ? new Map() : undefined; _wrapNativeSuper = function _wrapNativeSuper(Class) { if (Class === null || !_isNativeFunction(Class)) return Class; if (typeof Class !== "function") { throw new TypeError("Super expression must either be null or a function"); } if (typeof _cache !== "undefined") { if (_cache.has(Class)) return _cache.get(Class); _cache.set(Class, Wrapper); } function Wrapper() { return _construct(Class, arguments, _getPrototypeOf(this).constructor); } Wrapper.prototype = Object.create(Class.prototype, { constructor: { value: Wrapper, enumerable: false, writable: true, configurable: true } }); return _setPrototypeOf(Wrapper, Class); }; return _wrapNativeSuper(Class); }
function _construct(t, e, r) { if (_isNativeReflectConstruct()) return Reflect.construct.apply(null, arguments); var o = [null]; o.push.apply(o, e); var p = new (t.bind.apply(t, o))(); return r && _setPrototypeOf(p, r.prototype), p; }
function _isNativeReflectConstruct() { try { var t = !Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); } catch (t) {} return (_isNativeReflectConstruct = function _isNativeReflectConstruct() { return !!t; })(); }
function _isNativeFunction(fn) { try { return Function.toString.call(fn).indexOf("[native code]") !== -1; } catch (e) { return typeof fn === "function"; } }
function _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }
function _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }
var eventList = [];
var NonceElement = /*#__PURE__*/function (_HTMLElement) {
  _inherits(NonceElement, _HTMLElement);
  function NonceElement() {
    _classCallCheck(this, NonceElement);
    return _callSuper(this, NonceElement);
  }
  _createClass(NonceElement, [{
    key: "attributeChangedCallback",
    value: function attributeChangedCallback(name, oldValue, newValue) {
      eventList.push({
        type: "AttributeChanged",
        name: name,
        oldValue: oldValue,
        newValue: newValue
      });
    }
  }, {
    key: "connectedCallback",
    value: function connectedCallback() {
      eventList.push({
        type: "Connected"
      });
    }
  }], [{
    key: "observedAttributes",
    get: function get() {
      return ['nonce'];
    }
  }]);
  return NonceElement;
}( /*#__PURE__*/_wrapNativeSuper(HTMLElement));
customElements.define("nonce-element", NonceElement);
 </script>
 <nonce-element nonce="abc">
 </nonce-element>
 <script nonce="abc">
  "use strict";

test(function (t) {
  assert_object_equals(eventList[0], {
    type: "AttributeChanged",
    name: "nonce",
    oldValue: null,
    newValue: "abc"
  }, "AttributeChanged 1");
  assert_object_equals(eventList[1], {
    type: "Connected"
  }, "Connected");
  assert_object_equals(eventList[2], {
    type: "AttributeChanged",
    name: "nonce",
    oldValue: "abc",
    newValue: ""
  }, "AttributeChanged 2");
  assert_equals(eventList.length, 3);
}, "Custom elements expose the correct events.");
 </script>
 <!-- CSS Leakage -->
 <style>
  #cssTest { display: block; }
    #cssTest[nonce=abc] { background: url(/security/resources/abe.png); }
 </style>
 <script id="cssTest" nonce="abc">
  "use strict";

test(function (t) {
  var script = document.querySelector('#cssTest');
  t.add_cleanup(function () {
    return script.remove();
  });
  var style = getComputedStyle(script);
  assert_equals(style['display'], 'block');
  assert_equals(style['background-image'], 'none');
}, "Nonces don't leak via CSS side-channels.");
 </script>
</body>
